+++ st2client/st2client/models/datastore.py
@@ -27,3 +27,13 @@ class KeyValuePair(core.Resource):
     _plural = 'Keys'
     _plural_display_name = 'Key Value Pairs'
     _repr_attributes = ['name', 'value']
+
+    # Note: This is a temporary hack until we refactor client and make it support non id PKs
+
+    def get_id(self):
+        return self.name
+
+    def set_id(self, value):
+        self.name = value
+
+    id = property(get_id, set_id)