+++ st2common/st2common/exceptions/sensors.py
@@ -25,9 +25,5 @@ class TriggerTypeRegistrationException(SensorPluginException):
     pass
 
 
-class NoSensorsFoundException(StackStormBaseException):
-    pass
-
-
 class SensorNotFoundException(StackStormBaseException):
     pass