+++ st2common/st2common/transport/reactor.py
@@ -26,6 +26,7 @@ __all__ = [
 
     'TriggerDispatcher',
 
+    'get_sensor_cud_queue',
     'get_trigger_cud_queue',
     'get_trigger_instances_queue'
 ]
@@ -36,7 +37,19 @@ LOG = logging.getLogger(__name__)
 TRIGGER_CUD_XCHG = Exchange('st2.trigger', type='topic')
 
 # Exchange for TriggerInstance events
-TRIGGER_INSTANCES_XCHG = Exchange('st2.trigger_instances_dispatch', type='topic')
+TRIGGER_INSTANCE_XCHG = Exchange('st2.trigger_instances_dispatch', type='topic')
+
+# Exchane for Sensor CUD events
+SENSOR_CUD_XCHG = Exchange('st2.sensor', type='topic')
+
+
+class SensorCUDPublisher(publishers.CUDPublisher):
+    """
+    Publisher responsible for publishing Trigger model CUD events.
+    """
+
+    def __init__(self, url):
+        super(SensorCUDPublisher, self).__init__(url, SENSOR_CUD_XCHG)
 
 
 class TriggerCUDPublisher(publishers.CUDPublisher):
@@ -54,7 +67,7 @@ class TriggerInstancePublisher(object):
 
     def publish_trigger(self, payload=None, routing_key=None):
         # TODO: We should use trigger reference as a routing key
-        self._publisher.publish(payload, TRIGGER_INSTANCES_XCHG, routing_key)
+        self._publisher.publish(payload, TRIGGER_INSTANCE_XCHG, routing_key)
 
 
 class TriggerDispatcher(object):
@@ -93,4 +106,8 @@ def get_trigger_cud_queue(name, routing_key):
 
 
 def get_trigger_instances_queue(name, routing_key):
-    return Queue(name, TRIGGER_INSTANCES_XCHG, routing_key=routing_key)
+    return Queue(name, TRIGGER_INSTANCE_XCHG, routing_key=routing_key)
+
+
+def get_sensor_cud_queue(name, routing_key):
+    return Queue(name, SENSOR_CUD_XCHG, routing_key=routing_key)